Fix for dpkg error while installing pve-manager
While running the script, the following error will appear during the process:

proxmox-ve depends on pve-manager; however:
Package pve-manager is not configured yet.

This appears to be a combination of upstream issues and race conditions. Installing pve-manager by itself, prior to installing the proxmox-ve package, along with disabling optional dependencies, seems to fix this issue.

This pull request has the script install pve-manager early, before the other packages, so that it's already installed and configured by the time the remainder of the script performs the remaining functions.